% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dABoverlap.R
\name{afni_3dABoverlap}
\alias{afni_3dABoverlap}
\alias{ABoverlap}
\alias{aboverlap}
\title{AFNI 3dABoverlap function}
\usage{
afni_3dABoverlap(file1, file2, automask = TRUE)

ABoverlap(...)

aboverlap(...)
}
\arguments{
\item{file1}{nifti object or NIfTI filename for overlap}

\item{file2}{nifti object or NIfTI filename for overlap}

\item{automask}{if \code{TRUE}, \code{automask} is run on the files.
If \code{FALSE}, consider input datasets as masks 
(automask does not work on mask datasets)}

\item{...}{not used}
}
\value{
Output filename of the brik
}
\description{
Wrapper for AFNI \code{3dABoverlap} function
}
