% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_to_nifti.R
\name{afni_3dAFNItoNIFTI}
\alias{afni_3dAFNItoNIFTI}
\alias{afni_to_nifti}
\title{Convert AFNI File to NIFTI Image}
\usage{
afni_3dAFNItoNIFTI(file, outfile = NULL, opts = "", float = TRUE,
  retimg = TRUE, ...)

afni_to_nifti(...)
}
\arguments{
\item{file}{AFNI BRIK or HEAD file}

\item{outfile}{output filename or prefix.  The default is to always use
\code{.nii.gz} extensions and will override if \code{outfile} has a
\code{.nii} extension.}

\item{opts}{Additional options to pass to \code{3dAFNItoNIFTI}}

\item{float}{Should the \code{-float} option be called in 
\code{3dAFNItoNIFTI}}

\item{retimg}{Should an image be returned?}

\item{...}{additional arguments to \code{\link{readnii}}}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
the \code{outfile} argument is returned (or a temporary filename)
}
\description{
Converts an AFNI file to an NIfTI file or \code{nifti} 
using \code{3dAFNItoNIFTI}
}
