% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dAllineate.R
\name{afni_3dAllineate}
\alias{afni_3dAllineate}
\alias{allineate}
\title{AFNI 3dAllineate function}
\usage{
afni_3dAllineate(file, fixed, cost = c("hel", "ls", "mi", "crM", "nmi",
  "crA", "crU"), interpolator = c("linear", "cubic", "NN", "quintic"),
  final_interpolator = c("cubic", "linear", "NN", "wsinc5", "quintic"),
  typeOfTransform = c("affine_general", "shift_only", "shift_rotate",
  "shift_rotate_scale"), outfile = NULL, retimg = TRUE, opts = "",
  ...)

allineate(...)
}
\arguments{
\item{file}{nifti object or NIfTI filename to register to \code{fixed} image}

\item{fixed}{nifti object or NIfTI filename of fixed image, target of registration}

\item{cost}{cost function that defines the matching 
between the file and the fixed image 
ls= Least Squares [Pearson Correlation]

mi  = Mutual Information [H(b)+H(s)-H(b,s)]

crM = Correlation Ratio (Symmetrized*)

nmi = Normalized MI [H(b,s)/(H(b)+H(s))]

hel = Hellinger metric

crA = Correlation Ratio (Symmetrized+)

crU = Correlation Ratio (Unsym)}

\item{interpolator}{interpolation method to use during matching process
NN - nearest neighbor.}

\item{final_interpolator}{interpolation used to create the
output dataset. \code{wsinc5} - weighted sinc interpolation method}

\item{typeOfTransform}{transformation to use for registration 
shift_only - 3 parameters

shift_rotate - 6 parameters

shift_rotate_scale -  9 parameters

affine_general - 12 parameters}

\item{outfile}{Output filename (should not have an extension)}

\item{retimg}{Should an image be returned (\code{TRUE}) or a filename?}

\item{opts}{Additional options passed to \code{3dAllineate}}

\item{...}{additional arguments to \code{\link{afni_3dAFNItoNIFTI}}}
}
\value{
Output filename of the brik
}
\description{
Wrapper for AFNI \code{3dAllineate} function
}
