% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dAutomask.R
\name{afni_3dAutomask}
\alias{afni_3dAutomask}
\alias{Automask}
\alias{automask}
\title{AFNI 3dAutomask function}
\usage{
afni_3dAutomask(file, mfrac = 0.5, gradual_clipping = TRUE,
  n_peels = 1, n_neighbors = 17, outfile = NULL, retimg = TRUE,
  opts = "", ...)

Automask(...)

automask(...)
}
\arguments{
\item{file}{nifti object or NIfTI filename.  If more than one is given,
they are given names of the letters up to \code{z}.}

\item{mfrac}{The 'clip level fraction', which 
must be a number between 0.1 and 0.9, which is passed to 
\code{\link{afni_3dClipLevel}}}

\item{gradual_clipping}{The program uses a 'gradual' clip level by default,
To use a fixed clip level, set \code{gradual_clipping = FALSE}.}

\item{n_peels}{Peel the mask \code{n_peels} times, then unpeel.  Designed
to clip off protuberances less than 2*\code{n_peels} voxels 
thick.}

\item{n_neighbors}{Define the number of neighbors needed for a voxel
NOT to be peeled.  The 18 nearest neighbors in
the 3D lattice are used, so \code{n_neighbors} should be between
9 and 18.}

\item{outfile}{Output filename, should have a \code{.nii.gz}
extension}

\item{retimg}{Should an image be returned (\code{TRUE}) or a filename?}

\item{opts}{Additional options passed to \code{3dAutomask}}

\item{...}{additional arguments to \code{\link{afni_3dAFNItoNIFTI}}}
}
\value{
Output filename of the brik
}
\description{
Wrapper for AFNI \code{3dAutomask} function
}
