% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dClipLevel.R
\name{afni_3dClipLevel}
\alias{afni_3dClipLevel}
\alias{ClipLevel}
\title{AFNI 3dClipLevel function}
\usage{
afni_3dClipLevel(file, mfrac = 0.5, opts = "")

ClipLevel(...)
}
\arguments{
\item{file}{nifti object or NIfTI filename.  If more than one is given,
they are given names of the letters up to \code{z}.}

\item{mfrac}{Multiplier of the median for clipping (see notes)}

\item{opts}{Additional options passed to \code{3dClipLevel}}

\item{...}{not used}
}
\value{
Single character string of a nubmer 
(by default), unless \code{opts = "-doall"}
}
\description{
Estimates the value at which to clip the anatomical dataset using
\code{3dClipLevel} so that background regions are set to zero.
}
\note{
From AFNI docs:
Algorithm:

(a) Set some initial clip value using wizardry (AKA 'variance').

(b) Find the median of all positive values >= clip value.

(c) Set the clip value to \code{mfrac} of this median.

(d) Loop back to (b) until the clip value doesn't change.
}
