% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dFourier.R
\name{afni_3dFourier}
\alias{afni_3dFourier}
\title{AFNI 3dFourier function}
\usage{
afni_3dFourier(file, lowpass = NULL, highpass = NULL, ignore = NULL,
  retrend = FALSE, retimg = FALSE, opts = "", ...)
}
\arguments{
\item{file}{nifti object or NIfTI filename.}

\item{lowpass}{low pass filter with a cutoff in Hz}

\item{highpass}{high pass filter with a cutoff in Hz}

\item{ignore}{ignore the first \code{n} images}

\item{retrend}{Any mean and linear trend are removed 
before filtering. This will restore the trend after filtering.}

\item{retimg}{Should an image be returned (\code{TRUE}) 
or a filename?}

\item{opts}{Additional options passed to \code{3dFourier}}

\item{...}{additional arguments to \code{\link{readnii}}}
}
\value{
Output filename
}
\description{
Wrapper for AFNI \code{3dFourier} function
}
