% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dLME.R
\name{afni_3dLME}
\alias{afni_3dLME}
\title{AFNI 3dLME function}
\usage{
afni_3dLME(in_data_table, in_glt_table = NULL, in_glf_table = NULL,
  out_file, model, rand_eff = NULL, quant_vars = NULL,
  cor_str = "AR1", ss_type = 3, jobs = "all", extra_options = NULL)
}
\arguments{
\item{in_data_table}{dataframe with columns Subj, InputFile, and any other fixed (between) or random (within) variables}

\item{in_glt_table}{dataframe with columns n_glt, glt_label, and glt_code}

\item{in_glf_table}{dataframe with columns n_glf, glt_labef, and glf_code}

\item{out_file}{filename/path of output file}

\item{model}{predictors to model}

\item{rand_eff}{random effects to model}

\item{quant_vars}{specify quantitative variables}

\item{cor_str}{autocorrelation setting (default: AR1)}

\item{ss_type}{sums of squares type (default: 3)}

\item{jobs}{number of cores (default: all)}

\item{extra_options}{extra not common options}
}
\value{
out_file and afni_list, which contains the afni path, program, options, and command
}
\description{
Wrapper for AFNI \code{3dLME} function by Ekarin Pongpipat

3dLME is a AFNI Group Analysis Program with Linear Mixed-Effects Modeling
Approach created by Gang Chen.
https://afni.nimh.nih.gov/pub/dist/doc/program_help/3dLME.html
}
\examples{
in_glt_table <- read_csv("in_glt_table.csv")
in_glf_table <- read_csv("in_glf_table.csv")
in_data_table <- read_csv("in_data_table.csv")
afni_3dLME(in_data_table = in_data_table, in_glf_table = in_glf_table, out_file = "output.nii.gz", model = "cond*RT+age", quant_vars = "RT,age", rand_eff = "~1+RT")

}
