% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dQwarp.R
\name{afni_3dQwarp}
\alias{afni_3dQwarp}
\alias{Qwarp}
\alias{qwarp}
\title{AFNI 3dQwarp function}
\usage{
afni_3dQwarp(file, fixed, align = TRUE, align_opts = "",
  resample = FALSE, blur = 2.345, outfile = NULL, retimg = TRUE,
  opts = "", ...)

Qwarp(...)

qwarp(...)
}
\arguments{
\item{file}{nifti object or NIfTI filename to register to \code{fixed} image}

\item{fixed}{nifti object or NIfTI filename of fixed image, target of registration}

\item{align}{Should \code{\link{afni_3dAllineate}} be run for an affine alignment
done before warping?}

\item{align_opts}{Options for alignment, in a single string.}

\item{resample}{This option simply resamples the file dataset to match the
fixed dataset grid.  You can use this if the two datasets
overlap well (as seen in the AFNI GUI), but are not on the
same 3D grid.}

\item{blur}{Gaussian blur the input images (specified in FWHM!) voxels before
doing the alignment (the output dataset will not be blurred).
The default is 2.345 (for no good reason).}

\item{outfile}{Output filename (should not have an extension)}

\item{retimg}{Should an image be returned (\code{TRUE}) or a filename?}

\item{opts}{Additional options passed to \code{3dQwarp}}

\item{...}{additional arguments to \code{\link{afni_3dAFNItoNIFTI}}}
}
\value{
Output filename of the brik or nifti object, depending on \code{retimg}
}
\description{
Wrapper for AFNI \code{3dQwarp} function
}
