% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3drefit.R
\name{afni_3drefit}
\alias{afni_3drefit}
\alias{refit}
\title{AFNI 3drefit function}
\usage{
afni_3drefit(file, opts = "", ...)

refit(...)
}
\arguments{
\item{file}{nifti object or NIfTI filename to change the header}

\item{opts}{Additional options passed to \code{3drefit}}

\item{...}{not currently used}
}
\value{
Output filename of the image
}
\description{
Wrapper for AFNI \code{3drefit} function
}
