% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dSkullStrip.R
\name{afni_3dSkullStrip}
\alias{afni_3dSkullStrip}
\title{AFNI 3dSkullStrip function}
\usage{
afni_3dSkullStrip(file, opts = "", retimg = TRUE, ...)
}
\arguments{
\item{file}{nifti object or NIfTI filename.  If more than one is given,
they are given names of the letters up to \code{z}.}

\item{opts}{Additional options passed to \code{3dSkullStrip}}

\item{retimg}{Should a nifti be returned?}

\item{...}{additional arguments to \code{\link{afni_3dAFNItoNIFTI}}}
}
\value{
Output filename of the image or a \code{nifti} image, depending 
on \code{retimg}
}
\description{
Wrapper for AFNI \code{3dSkullStrip} function
}
