% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dTshift.R
\name{afni_3dTshift}
\alias{afni_3dTshift}
\alias{tshift}
\title{Convert AFNI File to NIFTI Image}
\usage{
afni_3dTshift(file, tr, tpattern = c("alt+z", "alt+z2", "alt-z", "alt-z2",
  "seq+z", "seq-z"), tzero = NULL, slice = NULL, ignore_vols = 0,
  opts = "", retimg = TRUE, ...)

tshift(...)
}
\arguments{
\item{file}{AFNI BRIK or HEAD file}

\item{tr}{Repetition time. You may attach the suffix 's' for seconds,
or 'ms' for milliseconds.}

\item{tpattern}{Pattern of slice time pattern.
alt+z = altplus   = alternating in the plus direction
alt+z2            = alternating, starting at slice #1 instead of #0
alt-z = altminus  = alternating in the minus direction
alt-z2            = alternating, starting at slice #nz-2 instead of #nz-1
seq+z = seqplus   = sequential in the plus direction
seq-z = seqminus  = sequential in the minus direction}

\item{opts}{Additional options to pass to \code{3dTshift}}

\item{retimg}{Should an image be returned?}

\item{...}{additional arguments to \code{\link{readnii}}}

\item{float}{Should the \code{-float} option be called in 
\code{3dTshift}}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
the \code{outfile} argument is returned (or a temporary filename)
}
\description{
Converts an AFNI file to an NIfTI file or \code{nifti} 
using \code{3dTshift}
}
