% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dTstat.R
\name{afni_3dTstat}
\alias{afni_3dTstat}
\alias{afni_3dtstat}
\alias{tstat}
\alias{Tstat}
\title{AFNI 3dTstat function}
\usage{
afni_3dTstat(file, retimg = FALSE, opts = "", ...)

afni_3dtstat(file, retimg = FALSE, opts = "", ...)

tstat(file, retimg = FALSE, opts = "", ...)

Tstat(file, retimg = FALSE, opts = "", ...)
}
\arguments{
\item{file}{nifti object or NIfTI filename.  If more than one is given,
they are given names of the letters up to \code{z}.}

\item{retimg}{Should an image be returned (\code{TRUE}) or a filename?}

\item{opts}{Additional options passed to \code{3dTstat}}

\item{...}{additional arguments to \code{\link{afni_3dAFNItoNIFTI}}}
}
\value{
Output filename
}
\description{
Wrapper for AFNI \code{3dTstat} function
}
