% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dUndump.R
\name{afni_3dUndump}
\alias{afni_3dUndump}
\title{AFNI 3dUndump function}
\usage{
afni_3dUndump(in_file_coord, in_file_ref = NULL, in_file_mask = NULL,
  sphere_radius = NULL, opts = NULL, out_file = NULL,
  verbose = FALSE)
}
\arguments{
\item{in_file_coord}{Input coordinate file}

\item{in_file_ref}{Input file that coordinate file will reference for coordinate space}

\item{in_file_mask}{Input file that will mask output file. specifically, sphere regions inside the mask will be set to 1 while sphere regions outside the mask will be set to 0}

\item{sphere_radius}{Radius of sphere in millimeteres}

\item{opts}{Additional options to passed into function}

\item{out_file}{Output file}

\item{verbose}{FALSE (default). If true, will print afni path and command}
}
\value{
Returns output of command execution
}
\description{
Wrapper for AFNI \code{3dUndump} function
}
