% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3dvolreg.R
\name{afni_3dvolreg}
\alias{afni_3dvolreg}
\alias{volreg}
\title{Perform Volume Registration for Motion Correction}
\usage{
afni_3dvolreg(file, interpolation = c("Fourier", "heptic", "quintic",
  "cubic", "linear"), rp_file = NULL, full_motion_file = NULL,
  opts = "", retimg = TRUE, ...)

volreg(...)
}
\arguments{
\item{file}{file or \code{nifti} object}

\item{interpolation}{interpolation done after realignment.}

\item{rp_file}{filename of motion parameter output, which should have columns:
roll  = rotation about the I-S axis (degrees)

pitch = rotation about the R-L axis,  (degrees) CCW

yaw   = rotation about the A-P axis  (degrees)

dS  = displacement in the Superior direction (mm)

dL  = displacement in the Left direction (mm)

dP  = displacement in the Posterior direction (mm)}

\item{full_motion_file}{filename of full motion parameter output, including fields
as in rp file, but also brik indices and root mean squares}

\item{opts}{Additional options to pass to \code{3dvolreg}}

\item{retimg}{Should an image be returned?}

\item{...}{additional arguments to \code{\link{readnii}}}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
the \code{outfile} argument is returned (or a temporary filename)
}
\description{
Performs Volume Registration for Motion Correction on 
an NIfTI file or \code{nifti} 
using \code{3dvolreg}
}
