% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dWinsor.R
\name{afni_3dWinsor}
\alias{afni_3dWinsor}
\alias{Winsor}
\alias{winsor}
\title{AFNI 3dWinsor function}
\usage{
afni_3dWinsor(file, radius = 1.5, trim = c(0.2, 0.8), repeat_filter = 1,
  keepzero = FALSE, lower_threshold = NULL, mask = NULL, opts = "",
  retimg = TRUE, ...)

Winsor(...)

winsor(...)
}
\arguments{
\item{file}{nifti object or NIfTI filename.  If more than one is given,
they are given names of the letters up to \code{z}.}

\item{radius}{include all points within 'distance' in the operation, where distance
is defined as sqrt(i*i+j*j+k*k), and (i,j,k) are voxel index offsets}

\item{trim}{Quantiles for the top and bottom of winsorize filter}

\item{repeat_filter}{repeat filter nn times.  If nn < 0, means to repeat filter until
less than abs(n) voxels change}

\item{keepzero}{don't filter voxels that are zero}

\item{lower_threshold}{set voxels at or below \code{lower_threshold} to zero}

\item{mask}{use a mask dataset - voxels NOT in the mask won't be filtered}

\item{opts}{Additional options passed to \code{3dWinsor}}

\item{retimg}{Should a nifti be returned?}

\item{...}{additional arguments to \code{\link{afni_3dAFNItoNIFTI}}}
}
\value{
Output filename of the image or a \code{nifti} image, depending 
on \code{retimg}
}
\description{
Wrapper for AFNI \code{3dWinsor} function, 
which applies a 3D 'Winsorizing' filter.
}
