% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anfi_dir.R
\name{afni_dir}
\alias{afni_dir}
\alias{afnidir}
\title{Get AFNI Directory}
\usage{
afni_dir()

afnidir()
}
\value{
Character path
}
\description{
Finds the AFNIDIR from system environment or 
\code{getOption("afni.path")}
for location of AFNI functions and returns it
}
