% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_edge_mask.R
\name{afni_edge_mask}
\alias{afni_edge_mask}
\title{AFNI Edge Mask}
\usage{
afni_edge_mask(file, ...)
}
\arguments{
\item{file}{nifti object or NIfTI filename to get edge mask, assumed binary.}

\item{...}{additional arguments passed to \code{\link{afni_3dcalc}}}
}
\value{
Output filename or nifti object (default)
}
\description{
Compute the edge voxels of a mask dataset.  An edge voxel is one    
that shares some face with a non-masked voxel.  This computation    
assumes file is a binary mas
}
