% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_quote_expression.R
\name{afni_quote_expression}
\alias{afni_quote_expression}
\title{Quote an Expression}
\usage{
afni_quote_expression(x, trim = TRUE)
}
\arguments{
\item{x}{character string}

\item{trim}{should \code{\link{trimws}} be run first}
}
\value{
A character string
}
\description{
Simple wrapper for ensuring a string has single quotes at the front
and end of the string
}
\examples{
afni_quote_expression("hey")
afni_quote_expression("'hey'")
afni_quote_expression(c("'hey'", "ho", ""))

}
