% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_subset_4d.R
\name{afni_subset_4d}
\alias{afni_subset_4d}
\title{Subset 4D images}
\usage{
afni_subset_4d(file, time_start = 0, time_end = "$", by = 1,
  retimg = FALSE)
}
\arguments{
\item{file}{4-dimensional Filename or nifti object to subset}

\item{time_start}{the time point to subset from. 
Remember timepoint numbering starts from 0}

\item{time_end}{the last time point to subset}

\item{by}{How should the subsetting be done.  If you want to 
subset every second time point, use by = 2 for example}

\item{retimg}{Should an image be returned (\code{TRUE}) 
or a filename?}
}
\value{
A \code{nifti} object or an output file depending on
\code{retimg}
}
\description{
Subset 4D images
}
\examples{
library(kirby21.fmri)
if (have_afni()) {
tdir = tempfile()
res = download_fmri_data(outdir = tdir)
file = kirby21.fmri::get_fmri_filenames(outdir = tdir)[1]
sub_file = afni_subset_4d(file, time_start = 0, time_end = 1)
dim(neurobase::readnii(sub_file))[4] == 2
}
}
