% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anfi_suffix.R
\name{afni_suffix}
\alias{afni_suffix}
\title{Get AFNI suffix for file}
\usage{
afni_suffix(x, default = "orig")
}
\arguments{
\item{x}{A vector of character filenames}

\item{default}{if no suffix is found, use this value.  You can also specify
\code{NA}}
}
\value{
A character vector, the same length as \code{x}, with suffixes
}
\description{
Returns
}
\examples{
afni_suffix("mybrik+orig.nii.gz")
afni_suffix("mybrik+orig[1].nii.gz")
}
