% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_tt_vent_mask.R
\name{afni_tt_vent_mask}
\alias{afni_tt_vent_mask}
\title{Create AFNI Ventricle Mask in Talairach space}
\usage{
afni_tt_vent_mask(dxyz = rep(2.5, 3), ...)
}
\arguments{
\item{dxyz}{Dimensions to Resample the Image}

\item{...}{additional arguments passed to \code{\link{afni_3dresample}}}
}
\value{
An object of class \code{nifti}, unless \code{retimg = FALSE} in
\code{...}
}
\description{
Example of making a ventricle mask, to be used if Freesurfer
is not available or needed before using afni_proc.py
}
\note{
See 
\url{https://afni.nimh.nih.gov/pub/dist/doc/program_help/afni_proc.py.html}
for example.  This code uses \code{TT_desai_dd_mpm+tlrc}
}
