% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_version.R
\name{afni_version}
\alias{afni_version}
\alias{afniversion}
\title{GET AFNI Version}
\usage{
afni_version()

afniversion()
}
\value{
Character vector of the version.
}
\description{
Finds the AFNI version from \code{anfni -ver}
}
\examples{
if (have_afni()) {
 afni_version()
}
}

