% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brik.stub.R
\name{brik.stub}
\alias{brik.stub}
\title{Grab filename without BRIK, HEAD or nii}
\usage{
brik.stub(x, bn = FALSE)
}
\arguments{
\item{x}{character vector of filenames ending in .nii or .nii.gz}

\item{bn}{Take \code{\link{basename}} of file?}
}
\value{
A character vector with the same length as \code{x}
}
\description{
Quick helper function to strip off .nii, .nii.gz, BRIK, BRIK.gz,
HEAD, HEAD.gz, or brik/head from filename
}
