% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkimg.R
\docType{methods}
\name{checkimg,afni-method}
\alias{checkimg,afni-method}
\title{Force object to filename}
\usage{
\S4method{checkimg}{afni}(file, allow_array = FALSE, ...)
}
\arguments{
\item{file}{character or \code{afni} object}

\item{allow_array}{allow arrays to be passed in}

\item{...}{options passed to \code{\link{afni_3dAFNItoNIFTI}}}
}
\description{
Ensures the output to be a character filename (or vector) from an input
image or \code{nifti/afni}.
}
