% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.afni.R
\name{get_afni}
\alias{get_afni}
\title{Create command declaring for the AFNI directory}
\usage{
get_afni(fallback = TRUE)
}
\arguments{
\item{fallback}{(logical) should \code{DYLD_FALLBACK_LIBRARY_PATH} be exported 
so that libs work?}
}
\value{
NULL if AFNI in path, or bash code for setting up AFNI DIR
}
\description{
Finds the AFNIDIR from system environment or \code{getOption("afni.path")}
for location of AFNI functions
}

