% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/have_afni.R
\name{have_afni}
\alias{have_afni}
\title{Logical check if AFNI is accessible}
\usage{
have_afni(...)
}
\arguments{
\item{...}{options to pass to \code{\link{get_afni}}}
}
\value{
Logical TRUE is AFNI is accessible, FALSE if not
}
\description{
Uses \code{get_afni} to check if AFNIDIR is accessible or the option
\code{afni.path} is set and returns logical
}
\examples{
have_afni()
}
