% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_to_afni.R
\name{nifti_to_afni}
\alias{nifti_to_afni}
\title{Convert NIfTI Image to AFNI FIle}
\usage{
nifti_to_afni(file, outfile = NULL)
}
\arguments{
\item{file}{NIfTI image or \code{nifti} object}

\item{outfile}{output filename or prefix.}
}
\value{
A character filename of the output BRIK
}
