% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dDespike.R
\name{afni_3dDespike}
\alias{afni_3dDespike}
\alias{despike}
\title{Convert AFNI File to NIFTI Image}
\usage{
afni_3dDespike(file, ignore_vols = 0, curve_order = NULL,
  cut_values = c(2.5, 4), mask = TRUE, dilate = 4, localedit = FALSE,
  method = c("OLD", "NEW"), opts = "", retimg = TRUE, ...)

despike(...)
}
\arguments{
\item{file}{AFNI BRIK or HEAD file}

\item{opts}{Additional options to pass to \code{3dDespike}}

\item{retimg}{Should an image be returned?}

\item{...}{additional arguments to \code{\link{readnii}}}

\item{outfile}{output filename or prefix.  The default is to always use
\code{.nii.gz} extensions and will override if \code{outfile} has a
\code{.nii} extension.}

\item{float}{Should the \code{-float} option be called in 
\code{3dDespike}}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
the \code{outfile} argument is returned (or a temporary filename)
}
\description{
Converts an AFNI file to an NIfTI file or \code{nifti} 
using \code{3dDespike}
}
