% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dresample.R
\name{afni_3dresample}
\alias{afni_3dresample}
\title{AFNI 3dresample function}
\usage{
afni_3dresample(file, bound_type = NULL, dxyz = NULL, orient = NULL,
  resample_mode = c("NN", "Li", "Cu", "Bk"), master = NULL,
  outfile = NULL, retimg = TRUE, opts = "", ...)
}
\arguments{
\item{file}{nifti object or NIfTI filename.}

\item{bound_type}{}

\item{dxyz}{}

\item{orient}{}

\item{resample_mode}{Resampling method. 
The options are for 'Nearest Neighbor', 'Linear', 'Cubic' 
and 'Blocky' interpolation, respectively.}

\item{master}{}

\item{outfile}{Output filename (should not have an extension)}

\item{retimg}{Should an image be returned (\code{TRUE}) or a filename?}

\item{opts}{Additional options passed to \code{3dresample}}

\item{...}{additional arguments to \code{\link{afni_3dAFNItoNIFTI}}}
}
\value{
Output filename of the brik
}
\description{
Wrapper for AFNI \code{3dresample} function
}
